requirements.txt
Django==2.0
pytz==2017.3
psycopg2==2.7.3.2
django-ipware==1.1.6
waitress==1.1.0
geoip2==2.6
ipython